Resolve some of pull request comments
