+++ frontend/src/util/types/Requisition.ts
@@ -2,7 +2,6 @@ import {User} from "./User";
 import {Project} from "./Project";
 import {Vendor} from "./Vendor";
 import {gql} from "@apollo/client";
-import {PresetColorType} from "antd/es/_util/colors";
 
 export type RequisitionStatus =
     "DRAFT" |
@@ -12,42 +11,6 @@ export type RequisitionStatus =
     "ORDERED" |
     "RECEIVED" |
     "CANCELLED";
-
-export const StatusToColor = (status: RequisitionStatus): PresetColorType => {
-    switch (status) {
-        case "DRAFT": return "magenta"
-        case "SUBMITTED": return "gold"
-        case "PENDING_CHANGES": return "orange"
-        case "READY_TO_ORDER": return "green"
-        case "ORDERED": return "blue"
-        case "RECEIVED": return "blue"
-        case "CANCELLED": return "red"
-    }
-}
-
-export const StatusToString = (status: RequisitionStatus): string => {
-    switch (status) {
-        case "DRAFT": return "Draft"
-        case "SUBMITTED": return "Submitted"
-        case "PENDING_CHANGES": return "Pending Changes"
-        case "READY_TO_ORDER": return "Ready to Order"
-        case "ORDERED": return "Ordered"
-        case "RECEIVED": return "Received"
-        case "CANCELLED": return "Cancelled"
-    }
-}
-
-export const StatusToStep = (status: RequisitionStatus): number => {
-    switch (status) {
-        case "DRAFT": return 0
-        case "SUBMITTED": return 1
-        case "PENDING_CHANGES": return 1
-        case "READY_TO_ORDER": return 2
-        case "ORDERED": return 3
-        case "RECEIVED": return 4
-        case "CANCELLED": return 1
-    }
-}
 
 export type Requisition = {
     id: number,
@@ -64,28 +27,6 @@ export type Requisition = {
     referenceString: string,
     canEdit: boolean,
     otherFees: number
-}
-
-// Sums up total costs for requisition items
-export const getTotalItemsCost = (requisition: Requisition): string => {
-    let total = 0;
-    for (const requisitionItem of requisition.requisitionitemSet) {
-        total += requisitionItem.quantity * requisitionItem.unitPrice
-    }
-
-    return '$' + total.toFixed(2);
-}
-
-// Sums up total costs for requisition items plus other fees
-export const getTotalCost = (requisition: Requisition): string => {
-    let total = 0;
-    for (const requisitionItem of requisition.requisitionitemSet) {
-        total += requisitionItem.quantity * requisitionItem.unitPrice
-    }
-
-    total += requisition.otherFees;
-
-    return '$' + total.toFixed(2)
 }
 
 export type RequisitionItem = {
