+++ frontend/src/util/util.ts
@@ -1,0 +1,48 @@
+import {Requisition, RequisitionStatus} from "./types/Requisition";
+import {PresetColorType} from "antd/es/_util/colors";
+
+export const StatusToColor = (status: RequisitionStatus): PresetColorType => {
+    switch (status) {
+        case "DRAFT": return "magenta"
+        case "SUBMITTED": return "gold"
+        case "PENDING_CHANGES": return "orange"
+        case "READY_TO_ORDER": return "green"
+        case "ORDERED": return "blue"
+        case "RECEIVED": return "blue"
+        case "CANCELLED": return "red"
+    }
+}
+
+export const StatusToString = (status: RequisitionStatus) => {
+    switch (status) {
+        case "DRAFT": return "Draft"
+        case "SUBMITTED": return "Submitted"
+        case "PENDING_CHANGES": return "Pending Changes"
+        case "READY_TO_ORDER": return "Ready to Order"
+        case "ORDERED": return "Ordered"
+        case "RECEIVED": return "Received"
+        case "CANCELLED": return "Cancelled"
+    }
+}
+
+export const StatusToStep = (status: RequisitionStatus) => {
+    switch (status) {
+        case "DRAFT": return 0
+        case "SUBMITTED": return 1
+        case "PENDING_CHANGES": return 1
+        case "READY_TO_ORDER": return 2
+        case "ORDERED": return 3
+        case "RECEIVED": return 4
+        case "CANCELLED": return 1
+    }
+}
+
+export const formatPrice = (num: number) => {
+    return '$' + num.toFixed(2);
+}
+
+export const getTotalCost = (requisition: Requisition, includeOtherFees: boolean) => {
+    let total = requisition.requisitionitemSet.map(item => item.quantity * item.unitPrice).reduce((prev, curr) => prev + curr, 0);
+
+    return includeOtherFees ? total + requisition.otherFees : total;
+}
